==> 01. Data insert.ipynb <==
import json
import ast
import pymongo
import pprint
import pandas as pd
from time import time
import ast
import sys
from pymongo import MongoClient

client = MongoClient("localhost", 27017)
#print(client.list_database_names())
db = client.CitationDB
col = db.papers

dbnames = client.list_database_names()
print(dbnames)

with open('dblp-ref-0.json', 'r', encoding='utf-8') as f:
    for line in f:
            # Convert line into json in 3 steps:
            # 1. json dumps converts to string,
            # 2. literal_eval takes string literal,
            # 3. json.loads converts to json
        data = json.loads(ast.literal_eval(json.dumps(line)))
        try:
                # Try inserting one line of data
            db.papers.insert_one(data)
        except TypeError as e:
                print(e)

with open('dblp-ref-1.json', 'r', encoding='utf-8') as f:
    for line in f:
            # Convert line into json in 3 steps:
            # 1. json dumps converts to string,
            # 2. literal_eval takes string literal,
            # 3. json.loads converts to json
        data = json.loads(ast.literal_eval(json.dumps(line)))
        try:
                # Try inserting one line of data
            db.papers.insert_one(data)
        except TypeError as e:
                print(e)

with open('dblp-ref-3.json', 'r', encoding='utf-8') as f:
    for line in f:
            # Convert line into json in 3 steps:
            # 1. json dumps converts to string,
            # 2. literal_eval takes string literal,
            # 3. json.loads converts to json
        data = json.loads(ast.literal_eval(json.dumps(line)))
        try:
                # Try inserting one line of data
            db.papers.insert_one(data)
        except TypeError as e:
                print(e)

query ={"year": {"$gte": "2005", "$lte":"2009"}}
#projections = {"id":1, "venue":1, 'n_citation': 1, "year":1}
journal_impact = list(db.col.find(query))

journal_impact
