==> 03. Dataframe.ipynb <==
import json
import pandas as pd
from pandas import json_normalize


# reading the JSON data using json.load()
file = 'focal_data.json'
with open(file) as focal_file:
    dict_focal = json.load(focal_file)


# converting json dataset from dictionary to dataframe
df = pd.DataFrame.from_dict(dict_focal, orient='columns')
#df.reset_index(level=0, inplace=True)

df2 = df.explode('references').reset_index(drop=True)

file = 'refer_data.json'
with open(file) as refer_file:
    dict_ref = json.load(refer_file)

# converting json dataset from dictionary to dataframe
df_ref = pd.DataFrame.from_dict(dict_ref, orient='columns')


df_ref.columns = ['ref_index', 'ref_abstract', 'ref_authors', 'ref_n_citation', 'ref_references', 'ref_title', 'ref_venue', 'ref_year', 'references']

df2['references'].astype(str)

df3 = pd.merge(df2, df_ref, on='references')

df3 = df3.astype({'abstract':str}) 
df3 = df3.astype({'ref_abstract': str})

df3

#similarity score as a new column 
from sklearn.feature_extraction.text import CountVectorizer
from scipy.spatial import distance
vectorizer = CountVectorizer()

vector_similarity =[]
for index, row in df3.iterrows():
    allsentences = [row['abstract'], row['ref_abstract']]
    all_sentences_to_vector = vectorizer.fit_transform(allsentences)
    text_to_vector_v1 = all_sentences_to_vector.toarray()[0].tolist()
    text_to_vector_v2 = all_sentences_to_vector.toarray()[1].tolist()
    cosine = distance.cosine(text_to_vector_v1, text_to_vector_v2)
    vector_similarity.append(round((1-cosine)*100,2))

df3['vector_similarity'] = vector_similarity


#Glove embedding
gloveFile = "glove.6B.50d.txt"
import numpy as np
def loadGloveModel(gloveFile):
    print ("Loading Glove Model")
    with open(gloveFile, encoding="utf8" ) as f:
        content = f.readlines()
    model = {}
    for line in content:
        splitLine = line.split()
        word = splitLine[0]
        embedding = np.array([float(val) for val in splitLine[1:]])
        model[word] = embedding
    print ("Done.",len(model)," words loaded!")
    return model

import re
from nltk.corpus import stopwords
import pandas as pd

def preprocess(raw_text):

    # keep only words
    letters_only_text = re.sub("[^a-zA-Z]", " ", raw_text)

    # convert to lower case and split 
    words = letters_only_text.lower().split()

    # remove stopwords
    stopword_set = set(stopwords.words("english"))
    cleaned_words = list(set([w for w in words if w not in stopword_set]))

    return cleaned_words

def cosine_distance_between_two_words(word1, word2):
    import scipy
    return (1- scipy.spatial.distance.cosine(model[word1], model[word2]))

def calculate_heat_matrix_for_two_sentences(s1,s2):
    s1 = preprocess(s1)
    s2 = preprocess(s2)
    result_list = [[cosine_distance_between_two_words(word1, word2) for word2 in s2] for word1 in s1]
    result_df = pd.DataFrame(result_list)
    result_df.columns = s2
    result_df.index = s1
    return result_df

def cosine_distance_wordembedding_method(s1, s2):
    import scipy
    vector_1 = np.mean([model[word] for word in preprocess(s1)],axis=0)
    vector_2 = np.mean([model[word] for word in preprocess(s2)],axis=0)
    cosine = scipy.spatial.distance.cosine(vector_1, vector_2)
    print('Word Embedding method with a cosine distance asses that our two sentences are similar to',round((1-cosine)*100,2),'%')


import scipy
model = loadGloveModel(gloveFile)


glove_similarity =[]
for i, r in df3.iterrows():
    s1 = r['abstract']
    s2 = r['ref_abstract']
    #print("--FOCAL--" + row['abstract'] + "--REFERENCE--" + row['ref_abstract'])
    s1 = preprocess(s1)
    s2 = preprocess(s2)
    for word1 in s1:
        vector_1 = np.mean(model[word1], axis=0)
    for word2 in s2:
        vector_2 = np.mean(model[word2], axis=0)
    cosine = scipy.spatial.distance.cosine(vector_1, vector_2)
    glove_similarity.append(round((1-cosine)*100,2))

#glove_similarity

#ADD NEW COLUMNS 1. NUMBER OF AUTHORS, 2. NUMBER OF REFERENCES, 3. SIMILARITY SCORES (2) 
#4. LATEST REFERENCED YEAR 5. MEDIAN REFERENCE YEAR 5. NUMBER OF CITATIONS OF REFERENCES

mean_similarity_count = pd.pivot_table(df3, values='vector_similarity', index=['id'], columns=None, aggfunc='mean')
max_similarity_count = pd.pivot_table(df3, values='vector_similarity', index=['id'], columns=None, aggfunc='max')

mean_similarity_count
max_similarity_count

df_main = pd.merge(df, mean_similarity_count,  on='id')

df_main = pd.merge(df_main, max_similarity_count,  on='id')

avg_ref_year = pd.pivot_table(df3, values='ref_year', index=['id'], columns=None, aggfunc='mean')

max_ref_year = pd.pivot_table(df3, values='ref_year', index=['id'], columns=None, aggfunc='max')

df_main = pd.merge(df_main, avg_ref_year ,  on='id')

df_main = pd.merge(df_main, max_ref_year ,  on='id')

df_main.rename(columns = {'vector_similarity_x':'mean_vector_similarity', 'vector_similarity_y':'max_vector_similarity',
                         'ref_year_x': 'mean_ref_year', 'ref_year_y': 'max_ref_year' }, inplace = True)

df3

mean_ref_citations = pd.pivot_table(df3, values='ref_n_citation', index=['id'], columns=None, aggfunc='mean')
df_main = pd.merge(df_main, mean_ref_citations,  on='id')

max_ref_citations = pd.pivot_table(df3, values='ref_n_citation', index=['id'], columns=None, aggfunc='max')
df_main = pd.merge(df_main, max_ref_citations,  on='id')

df_main.rename(columns = {'ref_n_citation_x':'mean_ref_n_citation', 'ref_n_citation_y':'max_ref_n_citation'}, inplace = True)

df_main

#add title word count as a column
title_len =[]
for index, row in df_main.iterrows():
    string = row['title']
    title_len.append(len(string.split()))

df_main['title_len'] = title_len

#add number of the authors
n_authors = []
for index, row in df_main.iterrows():
    liste = row['authors']
    n_authors.append(len(liste))

df_main['n_authors'] = n_authors

# n_authors, n_reference

n_references = pd.pivot_table(df3, values='references', index=['id'], columns=None, aggfunc='count')
df_main = pd.merge(df_main, n_references,  on='id')
df_main.rename(columns = {'references_y':'n_references'}, inplace = True)

df_main

# #of words in abstract (longer the more citations)

ab_words = []
for index, row in df_main.iterrows():
    string = row['abstract']
    ab_words.append(len(string.split()))

df_main['ab_len'] = ab_words

#journal impact factor ask if I can have an access to Journal Citation Reports (JCR) database

df_main


#save df3 as df_with_reference
df3.to_pickle("df_with_reference.pkl")
    

#save df_main as df_main
df_main.to_pickle("df_main.pkl")
