==> 04. Dataframe2.ipynb <==
import json
import pandas as pd
from pandas import json_normalize
import numpy as np
from pymongo import MongoClient
import pprint

client = MongoClient("localhost", 27017)
db = client["CitationDB"]
col = db.papers

dbnames = client.list_database_names()
print(dbnames)

db.newDB.count_documents({'year': {"$gte": "2005", "$lte":"2009"}})

query ={"year": {"$gte": "2005", "$lte":"2009"}}
#projections = {"id":1, "venue":1, 'n_citation': 1, "year":1}
journal_impact = list(db.col.find(query))


(journal_impact)

df_main = pd.read_pickle("df_main.pkl")
df_refer = pd.read_pickle("df_with_reference.pkl")

title_meta =[]
for index, row in df_main.iterrows():
    if "meta" in row['title']:
        title_meta.append(1)
    else:
        title_meta.append(0)

df_main['title_meta'] = title_meta

title_review =[]
for index, row in df_main.iterrows():
    if "review" in row['title']:
        title_review.append(1)
    else:
        title_review.append(0)

df_main['title_review'] = title_review

with open('journal_impact_1y.json') as json_file:
     impact_1y= json.load(json_file)

impact_1y.pop('') 

impact_1y['']=0

impact_1y_col =[]

for index, row in df_main.iterrows():
    if row['venue'] == "":
        impact_1y_col.append(0)
    else:
        if row['venue'] not in impact_1y.keys():
            impact_1y_col.append(0)
        else:
            for x, y in impact_1y.items():
                if row['venue'] == x:
                    impact_1y_col.append(y)
            

df_main['impact_1y'] = impact_1y_col

df_main['impact_1y'].describe()
#question: should I scale this column?

#save df3 as df_with_reference
df_main.to_pickle("df_main2.pkl")
