==> 05. refer_to_main.ipynb <==
import json
import pandas as pd
from pandas import json_normalize
import numpy as np
from pymongo import MongoClient
import pprint

df_main2 = pd.read_pickle("df_main2.pkl")
df_refer2 = pd.read_pickle("df_refer2.pkl")

df_main2.drop(['count_similarity_x','count_similarity_y', 'tfidf_similarity_x','tfidf_similarity_y'], inplace=True, axis=1)
df_main2.drop(['mean_vector_similarity','max_vector_similarity', 'mean_ref_n_citation', 'max_ref_n_citation'], inplace=True, axis=1)

pd.set_option("display.max_columns", None)

#df_refer2.drop(['w2v_Jac'], inplace=True, axis=1)
df_refer2['BoW_Euc'] = df_refer2['BoW_Euc']/100
df_refer2['w2v_Cos'] = df_refer2['w2v_Cos']*100

df_refer2.describe()

mean_similarity = pd.pivot_table(df_refer2, values=['BoW_Jac','BoW_Cos','BoW_Euc','tfidf_Jac', 'tfidf_Cos','tfidf_Euc','w2v_Cos'], 
                                                    index=['id'], columns=None, aggfunc='mean')
df_main3 = pd.merge(df_main2, mean_similarity,  on='id')

max_similarity = pd.pivot_table(df_refer2, values=['BoW_Jac','BoW_Cos','BoW_Euc','tfidf_Jac', 'tfidf_Cos','tfidf_Euc','w2v_Cos'], 
                                                    index=['id'], columns=None, aggfunc='max')
df_main3 = pd.merge(df_main3, max_similarity,  on='id')

min_similarity = pd.pivot_table(df_refer2, values=['BoW_Jac','BoW_Cos','BoW_Euc','tfidf_Jac', 'tfidf_Cos','tfidf_Euc','w2v_Cos'], 
                                                    index=['id'], columns=None, aggfunc='min')
df_main3 = pd.merge(df_main3, min_similarity,  on='id')

df_main3

df_refer2

df_main3.to_pickle("df_main3.pkl")
df_refer2.to_pickle("df_refer2.pkl")
